refractor(transaction): refractor all transaction flow, we have multiple transaction and single transaction

+++ lib/features/transaction/domain/entities/transaction_detail_request.dart
@@ -13,8 +13,8 @@ class TransactionDetailRequest {
 
   Map<String, dynamic> toJson() => {
         'scrapCategoryId': scrapCategoryId,
-        'pricePerUnit': pricePerUnit,
+        'pricePerUnit': pricePerUnit.toDouble(),
         'unit': unit,
-        'quantity': quantity,
+        'quantity': quantity.toDouble(),
       };
 }
